+++ src/auth/controller/index.ts
@@ -1,0 +1,16 @@
+import loginPage from './loginPage'
+import logout from './logout'
+import signUpPage from './signUpPage'
+import signUpProcess from './signUpProcess'
+import tryLogin from './tryLogin'
+
+const authController ={
+	loginPage,
+	tryLogin,
+	logout,
+  signUpPage,
+  signUpProcess
+}
+
+
+export default authController
