+++ src/auth/controller/tryLogin.ts
@@ -1,0 +1,6 @@
+import passport from 'passport'
+
+export default passport.authenticate('local', { 
+	successRedirect: '/recruitment',
+	failureRedirect: '/auth/login',
+ })
