+++ src/auth/router/authRouter.ts
@@ -1,12 +1,9 @@
 import express from 'express';
-import passport from 'passport';
-import authController from '../controller/authController'
+import authController from '../controller'
 
 
 const router : express.Router = express.Router({mergeParams : true});
-
 router.get('/login', authController.loginPage);
-
 router.post('/login',  authController.tryLogin, authController.loginPage)
 
 router.get('/logout', authController.logout)
