+++ src/myPage/controller/getMyPage.ts
@@ -1,0 +1,6 @@
+import { NextFunction, Request, Response } from 'express';
+
+export default (req : Request, res : Response, next : NextFunction) => {
+	
+	res.send('myPage');
+}
