+++ src/myPage/controller/index.ts
@@ -1,0 +1,9 @@
+import getMyPage from './getMyPage'
+
+
+
+const myPageController = {
+	getMyPage
+}
+
+export default myPageController;
