+++ src/myPage/router/myPageRouter.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+import myPageController from '../controller';
+import {isLogin} from '../../middleware/authCheck'
+
+const router: Router = Router();
+
+router.get('/', isLogin, myPageController.getMyPage);
+
+export default router;
