+++ src/recruitment/controller/RecruitmentController.ts
@@ -1,8 +1,8 @@
 import { NextFunction, Request, Response } from 'express';
-import authCheck from '../middleware/authCheck'
+
 
 const mainPage = (req : Request, res : Response) => {
-	if (authCheck.isLogin)
+	
 	res.send('default (all) Page!');
 }
