+++ src/recruitment/router/recruitmentRouter.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import RecruitmentController from '../controller/RecruitmentController';
-import AuthCheck from '../middleware/AuthCheck';
+import AuthCheck from '../../middleware/AuthCheck';
 
 const router: Router = Router();
 // jwt 토큰 있어야하는 것들 2번째 파라미터에 집어 넣으면 됨
